+++ src/GzipFile.h
@@ -11,7 +11,7 @@ class GzipFile : public BaseFileIO {
     ~GzipFile(); 
     int Open(const char *, const char *);    
     int Close();
-    long long int Size(char *);
+    off_t Size(char *);
     int Read(void *, size_t, size_t );
     int Write(void *, size_t, size_t);  
     int Seek(off_t);
